cm_data: Added code to synchronize at the end of each iteration via socket communication.
Refactored sockets communication code and moved it from complex folder to common folder for reuse,
added sychroniztion via sockets communication and added command line option "-q" to specify service port.
